Restore FS0023 for union Is* properties
Restores the error for a conflicting user-defined Is* property from:

error FS0438: Duplicate method. The method 'get_IsFoo' has the same name and signature as another method in type 'Foo'.

Back to:

error FS0023: The member 'IsFoo' can not be defined because the name 'IsFoo' clashes with the default augmentation of the union case 'Foo' in this type or module